+++ crypto/src/hash/rescue/rp64_256/digest.rs
@@ -68,6 +68,24 @@ impl Deserializable for ElementDigest {
     }
 }
 
+impl From<[BaseElement; DIGEST_SIZE]> for ElementDigest {
+    fn from(value: [BaseElement; DIGEST_SIZE]) -> Self {
+        Self(value)
+    }
+}
+
+impl From<ElementDigest> for [BaseElement; DIGEST_SIZE] {
+    fn from(value: ElementDigest) -> Self {
+        value.0
+    }
+}
+
+impl From<ElementDigest> for [u8; 32] {
+    fn from(value: ElementDigest) -> Self {
+        value.as_bytes()
+    }
+}
+
 // TESTS
 // ================================================================================================