+++ examples/src/rescue/rescue_63/mod.rs
@@ -9,17 +9,26 @@ use log::debug;
 use std::time::Instant;
 use winterfell::{
     math::{fields::f63::BaseElement, log2, FieldElement},
-    ProofOptions, StarkProof, VerifierError,
+    ProofOptions, Prover, StarkProof, Trace, TraceTable, VerifierError,
 };
 
 #[allow(clippy::module_inception)]
 pub mod rescue;
 
 mod air;
-use air::{build_trace, PublicInputs, RescueAir};
+use air::{PublicInputs, RescueAir};
+
+mod prover;
+use prover::RescueProver;
 
 #[cfg(test)]
 mod tests;
+
+// CONSTANTS
+// ================================================================================================
+
+const CYCLE_LENGTH: usize = 8;
+const NUM_HASH_ROUNDS: usize = 7;
 
 // RESCUE HASH CHAIN EXAMPLE
 // ================================================================================================
@@ -83,8 +92,12 @@ impl Example for RescueExample {
             ---------------------",
             self.chain_length
         );
+
+        let prover = RescueProver::new(self.options.clone());
+
+        // generate the execution trace
         let now = Instant::now();
-        let trace = build_trace(self.seed, self.chain_length);
+        let trace = prover.build_trace(self.seed, self.chain_length);
         let trace_length = trace.length();
         debug!(
             "Generated execution trace of {} registers and 2^{} steps in {} ms",
@@ -94,11 +107,7 @@ impl Example for RescueExample {
         );
 
         // generate the proof
-        let pub_inputs = PublicInputs {
-            seed: self.seed,
-            result: self.result,
-        };
-        winterfell::prove::<RescueAir>(trace, pub_inputs, self.options.clone()).unwrap()
+        prover.prove(trace).unwrap()
     }
 
     fn verify(&self, proof: StarkProof) -> Result<(), VerifierError> {